requirements.txt
aiofiles==25.1.0
aiosqlite==0.22.1
asyncua==1.1.8
cffi==2.0.0
cryptography==46.0.6
matplotlib==3.10.1
pytest==9.0.3
pytest-asyncio==1.3.0
pycparser==3.0
python-dotenv==1.0.1
pyOpenSSL==26.0.0
psutil==7.0.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
six==1.17.0
sortedcontainers==2.4.0
typing_extensions==4.15.0
wait_for2==0.3.2
